+++ cli/commands/analyze-optimization-detective-effectiveness.mjs
